drivers: power: supply: qcom: fix -Werror,-Wvoid-pointer-to-int-cast
../drivers/power/supply/qcom/oplus_qpnp_qg.c:5755:21: error: cast to smaller integer type 'u8' (aka 'unsigned char') from 'const void *' [-Werror,-Wvoid-pointer-to-int-cast]
 5755 |         chip->qg_version = (u8)of_device_get_match_data(&pdev->dev);
      |                            ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
1 error generated.

+++ drivers/power/supply/qcom/oplus_qpnp_qg.c
@@ -5752,7 +5752,7 @@ static int qpnp_qg_probe(struct platform_device *pdev)
 	chip->esr_actual = -EINVAL;
 	chip->esr_nominal = -EINVAL;
 	chip->batt_age_level = -EINVAL;
-	chip->qg_version = (u8)of_device_get_match_data(&pdev->dev);
+	chip->qg_version = (u8)(long)of_device_get_match_data(&pdev->dev);
 
 	switch (chip->qg_version) {
 	case QG_LITE:
